Don't force items to re-notify a change
Previously the property `items` had a simple observer `_updateItems`
that forced Polymer to re-evaluate the `items` array bindings
whenever its value was updated. This didn't make sense and led to
loops like this:

App sets `items` to new value (different reference)
-->
px-steps runs `_updateItems`, forces Polymer to notify the new
`items` value up to the app as if it was changed from inside
the component
-->
The app gets notified that the `items` value changed and treats it
as a new value even though its the same items it had before
-->
App sets `items` to a new-new value in px-steps and everything
repeats again

It does not make any sense to observe changes to the `items` property
and notify Polymer bindings that there were changes to the `items`
property. Polymer will only run `_updateItems` when there was a
notifying change in the first place. Removing this redundant code
since it seems to cause issues and doesn't solve any problems.